f-拆分model
