+++ dashboard/biz_routers/dataset.py
@@ -3,7 +3,7 @@ from jinja2 import TemplateNotFound
 from pydantic import BaseModel
 from starlette.requests import Request
 
-from dashboard.models import DataSet
+from dashboard.biz_models.dataset import DataSet
 from last.services.depends import create_checker, get_model_resource, get_resources
 from last.services.resources import Model as ModelResource
 from last.services.template import templates
